Revert "added extra constraint in the compilation. Now implicit casts are prohibited. Refactored all the code base accordingly. New bug are emerged still fixing"
This reverts commit 9bf5a6da3b513e885206f477bc29841a67ac4a0e.

+++ lib/raceup_board/components/can.h
@@ -4,7 +4,7 @@
 #include <stdint.h>
 
 typedef struct{
-  uint16_t id;
+  uint32_t id;
   union{
     uint8_t buffer[8];
     uint32_t words[2];
@@ -62,12 +62,12 @@ hardware_write_can(const struct CanNode* const restrict self ,
 extern struct CanMailbox*
 hardware_get_mailbox(struct CanNode* const restrict self, const enum MAILBOX_TYPE type,
     uint16_t filter_id, const uint16_t filter_mask,
-    const uint8_t mex_size)__attribute__((__nonnull__(1)));
+    const uint16_t mex_size)__attribute__((__nonnull__(1)));
 
 #define hardware_get_mailbox_single_mex(self,type,filter_id,mex_size)\
-  hardware_get_mailbox(self,type,filter_id,(uint16_t)(~0),mex_size)
+  hardware_get_mailbox(self,type,filter_id,~0,mex_size)
 
-extern int32_t
+extern int8_t
 hardware_mailbox_read(struct CanMailbox* const restrict self ,
     CanMessage* const restrict o_mex)__attribute__((__nonnull__(1,2)));
