+++ src/cores/core_0/feature/imu/imu.h
@@ -17,8 +17,8 @@ enum IMU_AXIS{
 
 int8_t imu_init(Imu_h* const restrict self )__attribute__((__nonnull__(1)));
 int8_t imu_update(Imu_h* const restrict self )__attribute__((__nonnull__(1)));
-float imu_get_acc(const Imu_h* const restrict self ,
+int8_t imu_get_acc(const Imu_h* const restrict self ,
         const enum IMU_AXIS axes)__attribute__((__nonnull__(1)));
-float imu_get_speed(const Imu_h* const restrict self )__attribute__((__nonnull__(1)));
+int8_t imu_get_speed(const Imu_h* const restrict self )__attribute__((__nonnull__(1)));
 
 #endif // !__GIEI_IMU__
